+++ api/v1/server/authn/middleware.go
@@ -252,7 +252,11 @@ func (a *AuthN) handleBearerAuth(c echo.Context) error {
 		// we permit exchange token auth if the token is valid and represents a user if the endpoint is not tenant-scoped, because
 		// this is effectively a PAT without the tenant scoping
 		if isTenantScoped && *tenantId != queriedTenant.ID {
-			a.l.Error().Msgf("tenant id in token does not match tenant id in context")
+			a.l.Error().
+				Str("tenant_id", tenantId.String()).
+				Str("queried_tenant_id", queriedTenant.ID.String()).
+				Str("user_id", userId.String()).
+				Msgf("tenant id in token does not match tenant id in context")
 
 			return forbidden
 		}